Move plugin code to its own file
For #252
